feat: Product Group Detail feature complete (Cobra Kai)
Phase 1 - Schema:
- product_group_sales table with DECIMAL qty column
- RLS policies and indexes

Phase 2 - Parser:
- Captures all 18 individual product groups on upload
- Saves to both product_mix_monthly and product_group_sales

Phase 3 - API:
- getProductGroupComparison with YoY growth calculation
- Period support: month, quarter, year, ytd
- Fixed YTD for past years (shows all 12 months)

Phase 4 - UI:
- PeriodSelector component with navigation
- ProductGroupChart (horizontal bar, prior year comparison)
- ProductGroupTable (mobile cards + desktop table)
- Detail page at /dealers/[id]/product-groups

Phase 5 - Backfill:
- Created backfill script for bulk CSV import
- Fixed qty INTEGER→DECIMAL type mismatch
- 2024 + 2025 data loaded

+++ supabase/migrations/015_add_product_group_sales.sql
@@ -11,7 +11,7 @@ CREATE TABLE product_group_sales (
   product_group TEXT NOT NULL,      -- e.g., 'NORTH STAR FLOORING'
   category TEXT NOT NULL,            -- e.g., 'ns_resp' (for filtering)
   sales DECIMAL(12,2) DEFAULT 0,
-  qty INTEGER DEFAULT 0,
+  qty DECIMAL(12,2) DEFAULT 0,
   orders INTEGER DEFAULT 0,
   period_start DATE,
   period_end DATE,
